requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
portia-sdk-python[google]>=0.7.2
PyGithub>=2.1.1
gitpython>=3.1.0
